Move java.net.http JSON-RPC client to consensusj-jsonrpc
* Move `org.consensusj.jsonrpc.javanet.JsonRpcClientJavaNet` to
  `org.consensusj.jsonrpc.JsonRpcClientJavaNet` in the `consensusj-jsonrpc` JAR.
* Remove the `consensusj-jsonrpc-javanet` module/JAR.
* Update documentation
