--- consensusj-jsonrpc-javanet/build.gradle
@@ -1,36 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-dependencies {
-    api project(':consensusj-jsonrpc')
-
-
-    testImplementation project(':cj-btc-jsonrpc')
-
-    testImplementation "org.apache.groovy:groovy:${groovyVersion}"
-    testImplementation ("org.apache.groovy:groovy-json:${groovyVersion}") {
-        transitive = false
-    }
-
-    testRuntimeOnly "org.slf4j:slf4j-jdk14:${slf4jVersion}"
-}
-
-ext.moduleName = 'org.consensusj.jsonrpc.javanet'
-
-jar {
-    inputs.property("moduleName", moduleName)
-    manifest {
-        attributes  'Implementation-Title': 'ConsensusJ JSON-RPC java.net.http client',
-                'Automatic-Module-Name': moduleName,
-                'Implementation-Version': archiveVersion.get()
-    }
-}
-
-test {
-    testLogging.showStandardStreams = true
-    systemProperty 'java.util.logging.config.file', "${project.projectDir}/src/test/logging.properties"
-    beforeTest { descriptor ->
-        logger.lifecycle('    ' + descriptor.getName())
-    }
-}
