+++ consensusj-jsonrpc/build.gradle
@@ -13,6 +13,8 @@ dependencies {
     // Add SLF4J runtime adapter for JDK logging for GraalVM native-image build of MathService/Tool
     nativeSampleImplementation "org.slf4j:slf4j-jdk14:${slf4jVersion}"
 
+    testImplementation project(':cj-btc-jsonrpc')
+
     testImplementation "org.apache.groovy:groovy:${groovyVersion}"
     testImplementation ("org.apache.groovy:groovy-json:${groovyVersion}") {
         transitive = false
